# Micro-optimize `binary_search`