We spend about 30% of cycles doing binary search in [`ExtendWith::count`](https://github.com/rust-lang/datafrog/blob/883f028f29436cd4df62f6a2f77b8c75b5f52235/src/treefrog.rs#L364-L371) while running `clap-rs/app-parser-{{impl}}-add_defaults`. This indicates to me that we need to explore different storage for the `cfg_edge` relation, but a maximally optimal `binary_search` is beneficial regardless.

This PR switches to `get_unchecked` to eliminate a bounds check that the optimizer cannot. This is also done in the [standard library implementation](https://github.com/rust-lang/rust/blob/507bff92fadf1f25a830da5065a5a87113345163/library/core/src/slice/mod.rs#L1930).

It also takes advantage of a lesser known invariant of `Vec`, that the capacity [cannot exceed `isize::MAX` bytes](https://doc.rust-lang.org/std/vec/struct.Vec.html#panics), to compute the midpoint using less instructions (there's a [fun article](https://ai.googleblog.com/2006/06/extra-extra-read-all-about-it-nearly.html) from the early internet about this). 

As a result, the aforementioned test case on a Ryzen 4700U laptop on a goes from this:

```
 Performance counter stats for 'target/release/polonius -a DatafrogOpt inputs/clap-rs/app-parser-{{impl}}-add_defaults':

   691,011,608,622      instructions:u                                              

      99.967872602 seconds time elapsed

      96.724862000 seconds user
       3.172634000 seconds sys
```

to this:

```
 Performance counter stats for 'target/release/polonius -a DatafrogOpt inputs/clap-rs/app-parser-{{impl}}-add_defaults':

   623,280,710,902      instructions:u                                              

      92.983953355 seconds time elapsed

      89.483132000 seconds user
       3.404928000 seconds sys
```

Wall time is highly variable and may differ on your platform.

cc @shepmaster (since they seem to be interested in this sort of thing)