@ecstatic-morse 2020-12-30
Also, we still use the standard library's binary search implementation in some places, which doesn't have the second (admittedly less important) optimization. If rust-lang/rust#74024 lands, we'll definitely want to switch to a custom impl in every case, since it will double the number of comparisons in the inner loop for little benefit.