@lqd 2021-02-28
We have talked about this PR on zulip, hopefully niko has some time for this (if not we will look at this at the next sprint).

I don't think bors is active on this repo, so let's ask manually via github 🙏